refactor(slack-bot): extract state.js from index.js (behavior-preserving)
